magit-branch-or-commit-at-point: check magit-buffer-refname
Return a meaningful value when we're visiting a blob.  This is
particularly convenient for exploring different files of a revision
because subsequent calls to magit-find-file will use the previously
selected revision as the default.

Revert the change from 96eee497 (magit-show-commit: use
magit-buffer-refname, 2016-08-08) because this case is now taken care
of by the magit-branch-or-commit-at-point call.

Re: #3066