git-commit-turn-on-auto-fill: Disable comment-auto-fill-only-comments
Buffers in text-mode do not define a comment syntax, which means that
auto-filling ignores a non-nil value for comment-auto-fill-only-comments
(i.e., auto-filling works on non-comments).  A user would reasonably
expect a commit buffer (in the default text-mode) to behave the same,
but it does not because git-commit-mode sets up the comment syntax,
resulting in comment-auto-fill-only-comments no longer being ignored.

Explicitly set comment-auto-fill-only-comments to nil to avoid this
discrepancy.

Fixes #3068.