magit-log-buffer-file: fix `-L` arg when region is active
Previously, the line numbers passed to `git log`'s `-L` arg were
calculated by subtracting 1 from `line-number-at-pos` for
`region-beginning` and `region-end`, which was incorrect.

e.g. calling `mark-whole-buffer` in a file with a single line with a
trailing newline would call `git log -L0,1:file`, instead of the
expected `git log -L1,1:file`.

This fix changes `magit-log-buffer-file` to calculate `-L` as a range
starting at the line number at the start of the region, and ending at
the line number at the end of the region (excluding the trailing
newline).