magit-completing-read: add default to choices if necessary
Until now, we did not bother adding the default choice to the list of
all choices, because we assumed that doing so was not necessary.

But it turned out that while it is possible to use such a default by
simply confirming it, `completing-read' does not actually add the
default to the list of all choices.  It therefore won't be shown when
the user uses TAB to list the choices.  And when REQUIRE-MATCH is
non-nil, then the user also won't be able to exit after manually
typing the default.

Additionally `ivy-completing-read' was recently changed to completely
ignore a default that isn't a member of the choices, making a fix more
pressing.

This commit simply prepends the default to the list of choices, if it
is not a `member' yet.  Doing this is much simpler than adjusting each
and every caller of `magit-completing-read', but it may in some cases
be weird for the default to be listed first, when the other choices
are sorted alphabetically.