magit-log-refresh-buffer: show line evolution range in header
Note that, even though "man git-log" says that the -L argument is
incompatible with a pathspec, we still consider -L when FILES is
non-nil.  We do so because, if the file from the -L argument and the
pathspec do *not* match, the output is empty, but git does not exit
with an error.  Showing information about both the pathspec and -L
argument in the header helps make the reason for the empty output more
obvious.  (When the file from the -L argument and pathspec do *not*
match, the output is the same as when no pathspec is given.)