magit-ref-ambiguous-p: new function
The primary purpose of magit-ref-fullname is to return a fully
qualified refname, but a few places in Magit use it to check whether a
refname is ambiguous (i.e., they don't use the fully qualified name
that's returned).  Add a new function, magit-ref-ambiguous-p, that
simply negates the output of magit-ref-fullname, making the intention
of the ambiguous checks more obvious.