magit-branch-merge-p: both checks must succeed, not just one of them
If a branch isn't merged into its upstream but it is merged into the
current branch, then `magit-branch-merged-p' used to consider it to be
merged.  But in the same situation Git does not consider the branch to
be merged.  Adjust this function to match Git's behavior.

Also remove the helper function `magit-branch-merged-into-upstream-p'.
`magit-branch-merged-p' used to be its only caller but now it doesn't
call it anymore.

Fixes #3107.