magit-rev-name: treat 'refs/<subdir>/*' as a namespace
Every 'git-name-rev --refs=PATTERN' instance in Magit has the
intention of restricting the results to refs within a particular
namespace ('refs/heads/', 'refs/remotes/', and so on).  This approach
has false positives, though, because git-name-rev matches PATTERN
against any part of the ref name, not only the beginning.  magit-wip's
names are frequent offenders (e.g., 'refs/wip/wtree/refs/heads/').

9ab6d554 (magit-rev-name: anchor the ref pattern by default,
2017-06-20) eliminated these non-namespace matches by automatically
tacking on an '--exclude=*/PATTERN', but we reverted it in b8d55867
because the --exclude option wasn't added until Git v2.13.0.

Instead let's discard git-name-rev's choice if it includes the
namespace as a substring.  Because the substring match could result
from a namespace repeated in the name (e.g., a terribly named
'refs/heads/refs/heads/foo'), we only discard the name if the ref
doesn't exist in the namespace.

A downside of this approach is that there may be some other ref in the
namespace that could be used to describe the revision, but we don't
have access to it because git-name-rev returns only its top
choice.  (The --exclude approach didn't have this downside.)  In the
magit-wip context, though, this situation appears to be uncommon.
When some commit can be named either as 'master~N' or
'wip/wtree/refs/heads/master~M', N is usually (always?) smaller than M
due to the way that wip refs are structured, so Git will choose
'master~N'.

Re: #3106, #3117