Support branch.<name.remote being an url instead of remote name
This fixes pushing to and pulling from the upstream when the current
branches `branch.<name.remote' is an url instead of a remote name as
is usually the case.  Also improve information shown on the "Merge:"
line in the status buffer.

It is possible that there are other places where the use or an url as
remote causes issues, but at least this fixes the issues mentioned in
issue #3116.