magit-commit-popup: add custom reader for --reuse-message
As mentioned in git-reset's manpage, ORIG_HEAD can be useful for
undoing a commit and reusing its message (though there are of course
other ways to do this).  Define a reader for the --reuse-message
option that offers ORIG_HEAD as the default value when it exists.

Also, prompt with all ref names.  If these aren't useful for the given
situation, the user can ignore them and enter an arbitrary value.

Closes #3110.