magit-show-commit: don't use log's file restriction with --follow
As of 1d8148dc, magit-show-commit and friends use the current log's
file restriction when they are called from a log buffer.  This
behavior, however, doesn't work well when the --follow option of 'git
log' is used because we don't know what the current file name is.

To avoid showing an empty diff for file renames in this situation,
ignore the log's file restriction when --follow is present in the
log's arguments.  Update the manual to mention this limitation and
point to the "-u" switch in the log popup as an alternative way to
view file-restricted diffs.