magit--insert-modules-logs: expect a range as argument
Previously it expected a function that returns one end of the range
(the other end always being the current branch).  Callers used to
pass either `magit-get-push-branch' or `magit-get-upstream-ref' as
that argument.

These functions have the advantage of returning nil if the respective
branch is not configured or does not exist, which in turn meant that
we did not end up calling `git log ...' only to learn that that log
is empty and then having to cancel the respective section.

However calling these functions is much more expensive than calling
`git log ...' and then possibly having to cancel.

For `@{upstream}' this change is un-problematic because that always
refers to exactly one branch or no branch at all.  The meaning of
`@{push}' however is overloaded and can refer to multiple branches,
or an unintended branch (for our use).

As far as I can tell (see `gitrevisions(7)'), the only variable that
affects this is `push.default', so to make sure `@{push}' works as
expected we have to use `git -c push.default=current log ...'.

Because of this issue the respective module section headings used to
mention "<push-remote>" but now the mention "@{push}", like other
headings mention "@{upstream}".  In this context ("show the one and
only push target of the current branch"), we force `git' to use a
meaningful value.  In other contexts, e.g. `git push' with no
arguments, that would be dangerous.  Here it is not, and it is also
not misleading because Magit never does `git push' with no arguments.
(It is possible to do so, but users have to explicitly opt-in to that
dangerous behavior.)