Add information about the fundraising campaign to status buffers
I am trying to make this as unobtrusive while still noticeable as
possible and to make it easy to permanently removed the section.

Some users might consider this to be an annoyance but it is also the
only way that can bring the campaign to the attention of a large
number of users who do not follow me on Twitter are did not visit
Reddit right after the launch.

I also hope that those users who are annoyed by this notice do weight
the brief interruption against the time they have saved by using Magit
and the time it has taken me to make that possible.

The notice will be shown from the 16th September to the end of the
1th October when the campaign ends.