git-rebase-mode-show-keybindings: detect translated output
Don't rely on the "Commands:" heading to detect the command section
because this can vary depending on $LANG.  Instead detect the line for
the first command (i.e. "p, pick").  There's no guarantee that this
output is stable but (1) the same is true for all interactive rebase
help output and (2) this text has not changed since the introduction
of interactive rebase.  Also, don't assume that each command and its
description are separated by "=" (e.g., "p, pick = use commit")
because some translations use a different separator.

The resulting output is a little odd in that it mixes English from
git-rebase-command-descriptions with Git's help in $LANG.  Another
option would be to force the entire rebase buffer to be in English by
setting LANG/LC_ALL (e.g., in magit-rebase-interactive-1).  Until we
decide that that's a direction we want to go, let's use this fix to
avoid displaying inaccurate bindings in rebase buffers with translated
text.

Fixes #3175.