git-commit-setup: pretend minor-modes are enabled earlier
Hook functions that are run by the major-mode specified by
`git-commit-major-mode' might want to consult the state of
the minor-modes `git-commit-mode' and `with-editor-mode', but
these minor-modes are not enabled until after the major-mode.
To pretend that the minor-modes are already enabled, bind the
minor-mode variables around the call to `normal-mode'.

Closes #3184.