magit-mode-map: Do bind <backtab>
Emacs is not consistent about the key produced by Shift+Tab across
different platforms, but <backtab> seems to be the common denominator.

On X Windows, Shift+Tab produced <S-iso-lefttab> which is translated
to <backtab>.  On terminal emulators such as Rxvt, <backtab> is
produced directly.

For w32 systems, S-<tab> is produced which is also translated to
<backtab>.

Therefore restore that binding for magit-section-cycle-global which
was earlier mistakenly removed with the unwanted Super+Tab binding.

Note that the Linux Virtual Terminal is unable to distinguish Tab from
Shift+Tab at all (Shift+Tab simply produces TAB).