Refuse to create instant fixup commit for an unreachable commit
When performing an instant fixup it is clearly a mistake to select an
unreachable commit.  The second step, the rebase, would always fail.

However creating a fixup commit for an unreachable commit without
trying to instantly rebase is a valid use-case, it allows "labeling"
the fixup as belonging to some commit.  Just doing a rebase would not
be enough to get the fixup to the right place, but we cannot know
whether the user is aware of that - so continue to allow the creation
of such fixup commits.