magit-blame-reverse: new command
The addition of this command makes it necessary to change the key
binding for `magit-blame-popup' in a few places, where "b" was used
before and "B" is used now.

Doing so allows moving through history using "b" and "f", if kept
`magit-blame-popup's old binding then we would have to use "b b" and
"f", which would be inconsistent.  Furthermore in `magit-blame-popup'
"b" was already bound to `magit-blame' and "B" to `magit-blame-popup,
which in turn is consistent with the other bindings in that popup.

It would be nice if we could rearrange `magit-file-popup' some more,
but due to limitations in `magit-popup' that is not possible at this
time.

Closes #3055.