magit-insert-{local,remote}-branches: use for-each-ref
9cac1f53 (magit-insert-{local,remote}-branches: use NUL separated
fields, 2017-10-11) breaks magit-show-refs because "git branch" does
not support the --format option in Git versions earlier than 2.13.

for-each-ref is almost as a drop-in replacement.  The only discrepancy
I've noticed is that the local refs do not include an entry for a
detached head.  At first glance, I don't see an issue with dropping
this because the buffer is about listing refs, and while there now no
longer is a line beginning with "@" when head is detached, the user
can still see in the header-line what commit the branches are being
compared too.

Fixes #3203.