magit-insert-local-branches: remove detached head check
In 2360d5b4, we switched to using "git for-each-ref --format ..."
instead of "git branch --format ..." for compatibility with Git
versions earlier than 2.13.  The only visible difference in the output
is that, when HEAD is detached, there is no longer a "(HEAD detached
at ...)" line.  Remove the leftover check for this line.