magit-format-remote*tagOpt: account for remote's length
The "remote.<name>.tagOpts" variable is not aligned with the other
variables in the remote popup unless the remote is 6 characters (which
is unsurprisingly the length of "origin").  Beyond the cosmetic
alignment issue, calling magit-remote-popup when the current branch's
remote is over 7 characters triggers an error because
magit-popup-format-variable passes a negative length argument to
make-string.

Fix the above issues by calculating the WIDTH argument based on the
remote's length.