Always show recent commits in the status buffer
Previously recent commits were only shown when they had not been pushed
to the upstream yet.  `magit-insert-unpulled-from-upstream-or-recent'
allowed showing recent commits more often, but it is strange that this
functions behavior depended on whether there were unPULLED commits.

Replace that function by one that falls back to displaying recent
commits if there are not unPUSHED commits instead and use that in the
default value of `magit-status-sections-hook'.  To avoid breakage
define an alias for the old function.