magit-log-refresh-buffer: add --full-history for single removed file
To show the history of removed files one has to use `--full-history'.
When showing the history for a single such file use that argument,
else not a single commit would be shown.  This can also results in
many unrelated merge commits being shown, which is why we don't use
this argument all the time, but in this special case that noise is
still preferable to nothing.

Closes #3012.