magit-visit-ref: drop prefix before calling magit-show-commit
Unless `focus-on-ref' is a member of `magit-visit-ref-behavior', the
prefix argument was simply passed along to `magit-show-commit', which
was inconvenient because that then prompted for the commit to show
instead of just showing the commit at point (tag or tip of branch).

Now we drop the prefix argument before calling `magit-show-commit',
but only if that command would not have been called without a prefix
argument anyway.

Closes #3009.