magit-visit-ref: offer plain checkout in addition to reset
When `create-branch' is a member of `magit-visit-ref-behavior' then
"visiting" a remote branch causes the creation and checkout of a local
branch by the same name.  If the local branch already exists, then the
user was offered to reset it to the current tip of the remote branch.

Now the user can also choose to check it out as-is, and when both tips
are the same commit, then the local branch is checked out without the
user having to confirm that.

Closes #3009.