magit-diff-highlight-list: fix invalid-selection special case
When the region was active but was not valid as a magit-selection
while beginning and/or ending inside a list section, then the heading
and children before the beginning of the region were not properly
highlighted to indicate that the list section is the current section.

This issue was introduced in 339e4f01c, which did not consider this
special case of the special case it addressed.

Re #3238.