magit-revert-popup: default to edit message on all platforms
`git-revert' defaults to let the user edit the commit message if and
only if it is run "from a terminal".  Magit uses a pty in most cases
but not on Cygwin, so here the behavior differs between platforms.

Add the `--edit' argument to the popup and enable it by default for
consistent behavior.  Also add the `--no-edit' argument.  It would be
better to have a single `--[no-]edit' argument, but the current popup
implementation does not support that.

Closes #3239.