magit-diff-while-committing: handle visible diff buffer separately
Replace the "diff buffer is current" state with the "diff buffer
is visible state".  Here "visible" means "being displayed in the
current frame.  If it is being displayed in another frame, then
we cannot tell whether the user can actually see it, because that
depends to much on the window manager.