magit-commit-diff: never give up
Fallback to the same base logic as `magit-diff-while-committing' to
decide what diff to display instead of giving up when we don't know
what diff to display based on `last-command'.

We cannot use that function, because if we did then we might end up
toggling to the "other diff" if the diff buffer already exists.

The "unrecord" and "abort" logic have be moved to the "outside".