magit-diff-while-committing: only toggle if --cached is in effect
Previously we did make sure that the staged changes were being
displayed before "toggling" to showing the amend changes.  Likewise
we did not make sure that the amend changes were being displayed
before "toggling" to display the staged changes.

Now the first to cases of the `cond' take care of toggling.  If
neither case matches, then we use the same logic that is used when
the diff buffer does not exist.

Also remove the distinction between "buffer exists and is visible"
and "buffer exists but is not visible".  The latter used some weird
logic that only made sense because `magit-commit-diff' did in some
cases not display the diff buffer upfront.  But since the previous
commit it always does.

Now we do this:
- If the most appropriate diff is not already being displayed,
  then display it.
- If the most appropriate diff is already being displayed,
  then toggle to the other.