magit-completion-read: add FALLBACK argument
This argument is only used if DEF is nil and its use bypasses the new
mechanism that allows users to skip confirmation.  It is useful when
it is safe to use the primary default without a chance to abort, but
not to do so for some secondary fallback default.

This will be used by `magit-read-stash', which already does something
similar.