magit-stash-{drop,pop}: ask for confirmation
`magit-stash-drop' did previously offer to act on another stash
instead, which some users, including myself, found inconvenient.
This was discussed in #2916.

`magit-stash-pop' did pop the stash without requiring confirmation,
which is unsafe because users might not expect "A" to do something
different on a stash than elsewhere.  This was discussed in #3220.