Add support for not showing certain errors to user
Instead of a single regexp try a list of regexps in order until one
matches the output.  If that regexp has a first submatch, then use
that as the error message as before.  If it matches but does not,
have a first submatch, then don't show a message (later regexps are
not tried).  When `magit-process-raise-error' is non-nil, then try
all regexps until one has a first submatch.

Do advertise this to users by using option because there probably
only is a limited set of errors that should be suppressed.