Don't use magit-graph face for summary
`ansi-color-apply-on-region' is currently used on the complete output.
It would be better to just apply the control characters on the graph
part of each line.  That's the only part that contains such characters
anyway.

Because we don't do that and because we want to use the `magit-graph'
for the "*" in the graph and currently that means using that face for
everything that doesn't already have a color.

Apparently nobody ever customized that face (otherwise this bug would
have been reported), so we could just remove it and stop colorizing
the graph stars along with the non-graph parts.

This commit is a compromise.  We just fallback to remove the face
later on for parts that should not use it.

Beside from the summary, we also remove it from the refs.  The latter
doesn't actually matter as far as users are concerned because here the
wrong face is later replaced by another.  But not removing the wrong
face first makes that task of adding the right face confusing.  I
noticed this bug while working on the respective code.