README.md: remove bountysource to avoid advertising it
I still accept donations there but I do not want to advertise it and
while I won't do so for the time being it is quite possible that I
will eventually encourage users to move their donations away from this
abysmal service.

The owners take a 10% share (on top of the processing fees), while the
work is being done by overworked volunteers.  Support requests are not
being answered, unless it happens in public (on Twitter).  Payouts
aren't performed by the end of the week as promised but take a month
or more.  It has been suggest that at least the documentation should
be adjusted accordingly, but nothing has happened.  Once your login
has timed out it is not possible to login again (or doing so would at
least require some research on the users part) and the underlying
issues have been reported three months ago.

Did I mention you pay 10% for this?