refs: preserve cherry-less, expanded sections on refresh
Expanding a cherry-less section results in an empty body.  Normally,
during refresh, an empty body causes the section to be removed by
`magit-cancel-section' (in this case called by `magit-git-wash'), but
because inserting the body is delayed in this case (for performance
reasons) until it is expanded `magit-insert-section--current' is nil
at the time this happens, which causes `magit-cancel-section' to do
nothing.

That's actually what we want, but that wasn't by design.  However when
the user then refreshes the buffer, the section remains expanded and
because it is being recreated during a refresh `magit-cancel-section'
does its job.

Prevent that by binding `magit-insert-section--current' to nil while
calling `magit-git-wash'.

Fixes #3267.