Properly deal with merges when cherry-picking
Change all commands which cherry-pick and revert commits to use the
new function `magit--cherry-pick'.  Previously all of these functions
used `magit-assert-one-parent' to deal with merges among the commits
to be applied.  But that function wasn't adequate and this commit
removes it.

It is not possible to apply merges and non-merges at once, because
`--mainline' must be provided for merges while its use causes an error
for non-merges.  In that case aborting is the only choice.

But otherwise add or remove the `--mainline' as appropriate.  When
`--mainline' is added, then ask the user for the parent number, when
it is removed, then do so without prompting.  This makes it possible
for users to use `--mainline=1' (which is the most reasonable choice
for most merges) by default without getting an error when applying
non-merges.

Closes #2980.