[DUMP] magit-diff-visit-file: visit HEAD on uncommitted removals
Since 050c8f070f this command visits the last commit that still had
that line when invoked with point on a removed line in a diff about a
committed change.

That commit also mentioned that we would do the same for unstaged and
staged changes and this commit finally implements that.

Setting `magit-diff-visit-previous-blob' to t reverts this change, but
I think the new behavior is very useful as it makes it much faster to
check why some line that is about to be removed was originally added:
"RET" to visit the blob from the commit that still had it (always
"HEAD" in this case) and then "b b" to find the commit that introduced
it.