git-commit-turn-on-flyspell: avoid unnecessary work
Our `git-commit-flyspell-verify' does its job as intended, but that
only keeps flyspell from actually checking words inside comments,
it does not keep it from looking at each word in comments and from
incorrectly claiming that it is checking them.

That takes time and is confusing.  To avoid it, call `flyspell-region'
on the relevant part of the buffer only.

We cannot use `forward-comment' because that is syntax based, but
we cannot use syntax for comments because the comment character only
begins a comment if it is located at the beginning of a line.