git-commit-font-lock-keywords-2: replace ineffective element
In the past Git might have said:
  Not currently on any branch.
instead of that it now says:
  HEAD detached at abcdefg.

`git-commit-font-lock-keywords-2' used to have an entry for the former
but it did not have any effect because only the first submatch would
have been colorized but the regexp had no submatches.

As a result the face `git-commit-comment-detached' wasn't actually
ever used.  Now it is used for the "HEAD" in the new output.  Also
remove the old bogus entry and don't bother colorizing anything for
older Git versions.