Enable "repository-local" values to be cached and retrieved
This cache should be used instead of buffer-local variables in
situations where the buffers in question are not guaranteed to still
exist at the point when the value is needed.

The following functions are provided:

* (magit-repository-local-repository)
* (magit-repository-local-set KEY VALUE &optional REPOSITORY)
* (magit-repository-local-exists-p KEY &optional REPOSITORY)
* (magit-repository-local-get KEY &optional DEFAULT REPOSITORY)
* (magit-repository-local-delete KEY &optional REPOSITORY)

KEY arguments should be symbols, or otherwise comparable by `eq'.

`magit-repository-local-repository' returns the repository identifier
for the current buffer.  In cases where asynchronous or other
time-delayed evaluation could make the "current buffer" indeterminate
at the point when subsequent repository-local functions are called,
this identifier should be captured and passed to those functions as an
explicit REPOSITORY argument.