Display up to 20 lines of the process error in the mode line tooltip
Rather than just repeating the single-line error summary (which is
frequently the fallback value "Git failed"), a tooltip can usefully
display multiple lines of text, so we can use it to show the actual
text from the process log.

As this output might be arbitrarily large, we constrain it to a
maximum number of lines.