git-commit.el: only require diff-mode when needed
Unlike what I said in the previous commit `diff-mode' does not
normally get loaded anyway.  So that commit resulted in it being
loaded needlessly in the normal case.

Instead just load it when we are about to use it and declare the
used variable so that the byte-compiler knows that the variable
uses dynamic scope and has nothing to complain about.