magit-maybe-save-repository-buffers: avoid redisplaying same message
We redisplay the previous message after displaying messages about
buffers being saved because that message is more important and would
otherwise not be displayed long enough for the user to see it.

But because `current-message` may return nil we ended up redisplaying
that message even if no message about buffers being saved were
displayed.

Fixes #3319.