Move Github-supported documentation files into sub-directory
Unfortunately Github only finds these files if they are located at the
top-level or in a sub-directory named either "docs" or ".github".  Our
documentation directory is called "Documentation".

We used to place these files at the top-level but since I am going to
add more of the supported files this would become to noisy.  The best
work around would be to create a symlink ".github -> Documentation,
but Github doesn't support that.  The next best thing is to create a
directory named ".github" that contains symlinks to files inside
"Documentation".