Inhibit magit-repository-local-set when the repository is unknown
Necessary now that `magit-repository-local-repository' returns nil
(rather than throwing an error) when no current repository can be
established.  It would be dangerous to allow values to be cached for
the nil repository, as there's no guarantee that such cached values
would only be retrieved in the intended context.

If it is necessary to cache a value in such circumstances, then an
explicit REPOSITORY argument should be passed.