Rename magit-{submodule-fetch => fetch-modules} and move definition
Now that this command boils down to "git fetch --recurse-submodules"
it makes more sense to (1) define it along with other fetch commands
and (2) invoke it from the fetch popup instead of the modules popup.

For the time being I keep the binding in the modules popup, but that
might change.  I would even argue that this command should be removed
completely and be replaced by adding "--recurse-submodules" to the
fetch popup.  But I suspect that doing so would upset muscle-memory,
so I am delaying that until I am absolutely certain.