magit-confirm: Fix regression
Unfortunately doing so means changing the function signature in a way
that is not backward compatible.

This function relies on being able to tell whether ITEMS is nil
because that argument was not provided or because an empty list was
provided as its value.  In the former case the action still has to be
confirmed in the latter case it does not because there is nothing to
act on.

So the `(items nil sitems)' argument has to be last but in d6420549c
I added a new argument NOCONFIRM after it.

Fixes #3341.