magit-worktree-delete: check for worktree directory before deleting
'git worktree list' (and thus magit-list-worktrees) may return
worktrees for which a directory no longer exists (e.g., if the
directory was deleted outside of magit).  Avoid the delete-directory
call in this case because it leads to an error and because the "git
worktree prune" call that follows is sufficient cleanup.