Don't pass read-file-name's result directly to git
When Magit is used with Tramp, read-file-name will return a path like
/ssh:host:/path/to/file.  The host part needs to be stripped before
passing the path to the host's git.  bf4332c8 fixed this issue for
magit-am-apply-{patches,maildir}.  This commit should take care of the
remaining places that call git with an unconverted read-file-name
result.

Follow the approach from bf4332c8 of converting a path at the time it
is passed to the git function rather than when it is returned by
read-file-name.  Note, however, that magit-file-rename is the only
function from bf4332c8 or this commit that makes use of the full
host:path, so it might make sense to add a function that wraps
(magit-read-file-name-convert (read-file-name ... )).

Re: #3368