magit-read-pull-request: Fix two bugs
Always request the individual pr even if it was part of the list
because the latter lacks some information.  That fixes #3371.

Previously we only tried to do so when the pr did not appear in
the list, but that didn't work because the use api resource was
wrong.  Fix that too.