magit-buffer-name-format: put * at both ends or neither
As per the conventions, put stars at both ends of buffer names if
possible.  When using the `uniquify' package that is not possible
because then the name must end with the path.  Previously we put
a star at the beginning anyway, but that is ugly so we no longer
do that.