whitespace-dont-turn-on-in-magit-mode: and maybe don't raise error
I though it was obvious that the old message implied "... therefore
do not enable Whitespace mode in Magit buffers.  Please adjust your
Whitespace mode configuration to prevent that from happening."  But
it turned out that is not the case.

Change the advice to silently prevent the activation if the attempt
is made due to the use of `global-whitespace-mode', but if the user
attempts to enable the local minor-mode directly, then show an error.

The hope is that this compromise results in the fewest invalid bug
reports.  (There is no bug - not in Magit at least.)

Closes #3374.