magit-branch-commit-at-point: Prefer branch over tag
Previously if point was on a revision in a lot and both a branch and
a tag point at that commit, then tag was returned.  But this function
should return a branch whenever possible and only fall back to a tag
as a slightly nicer alternative than just a hash.

One consequence of the old behavior was that `magit-checkout' often
ended up detaching HEAD instead of checking out a branch.