magit-confirm: List targets at bottom of frame
Also make it more obvious that the list is "part of the question" by
not separating the list and the question with a mode-line, but make
sure that there is at least one empty line between the list and the
prompt. Technically that line may actually be part of the prompt.