magit-branch-pull-request: Use numbered branch in case of conflict
Since 4f7f55c93 we check whether the maintainer is allowed to push
to the fork and if so create a local branch whose name matches that
of the branch on the fork.

When the contributor did not bother to create a feature branch, then
that won't work, because in that case master (or whatever the default
branch is called) already exists.  Since the mentioned commit this
resulted in a (appropriate) error.

Continue to check whether we are allowed to push to the fork, but if
the branch name conflicts then never-the-less create a branch "pr-N"
instead of attempting to create one named "master", just as if we were
not allowed to push to "fork/master".

Also set the upstream repository, usually "origin", as the push remote
even though we actually could push to "fork/master".  Git's push
remote mechanism only allows specifying the remote but the name of the
push-target branch on the remote is always the same as that of the
local branch). We cannot configure "pr-N" to be pushed to
"fork/master" (Git doesn't support it) and we cannot push to
"fork/pr-N" (we don't have the necessary permissions), so we configure
"origin" as the push-remote and therefore push to "origin/pr-N".